// src/types/ProjectData.ts
// Generated by https://quicktype.io

export interface ProjectData {
  id:         number;
  imagePath:  string;
  imageList:  string[];
  vidoPath:   string;
  name:       string;
  type:       Type;
  deskripsi:  string;
  tech:       string;
  linkApk:    string;
  linkGithub: string;
  linkWeb:    string;
  framework:  string[];
}

export enum Type {
  Mobile = "mobile",
  Web = "web",
}